[CAN-43] feat: THEME_COLORS, THEMES 컴포넌트에서 선언, constants/theme.ts 삭제
